COMP: Fix Python wrapping with VTK 7.1
When Slicer is built with the VTK 7.1, Python wrapping failed with errors
similar to:

    C:\path\to\Slicer-build\Modules\Remote\EMSegment\Qt\Logic\vtkTimeDefPython.cxx(73): error C2440: 'return' : cannot convert from 'vtkTimeDef *' to 'vtkObjectBase *'
              Types pointed to are unrelated; conversion requires reinterpret_cast, C-style cast or function-style cast
    C:\path\to\Slicer-build\Modules\Remote\EMSegment\Qt\Logic\EMLocalShapeCostFunctionPython.cxx(24): error C2440: 'static_cast' : cannot convert from 'vtkObjectBase *' to 'EMLocalShapeCostFunction *'
              Types pointed to are unrelated; conversion requires reinterpret_cast, C-style cast or function-style cast

This happens because VTK wrapping no longer recognizes BTX/ETX markers as
indicators to skip wrapping of a block of code; see Kitware/VTK@55878a2. In many
cases the updated wrapping tools can handle code that was previously had to be
skipped, but in this case wrapping still fails.

In VTK 7.1 blocks of code can be skipped for wrapping by checking whether
__VTK_WRAP__ is defined.



git-svn-id: http://svn.slicer.org/Slicer3/branches/Slicer4-EMSegment@17120 3bd1e089-480b-0410-8dfb-8563597acbee